requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
openai>=1.12.0
python-dotenv>=1.0.0
markdown>=3.5.0
pyperclip>=1.8.2
